+++ src/App.jsx
@@ -1,9 +1,9 @@
-
 import './App.css'
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom"
 import LayoutAdmin from './layoutAdmin/layoutAdmin'
-import LoginAdmin from './comps_admin/loginAdmin'
 import UsersList from './comps_admin/users/usersList'
+import FoodsList from './comps_admin/foods/foodsList'
+import Login from './assets/loginPage/login'
 
 function App() {
 
@@ -12,13 +12,34 @@ function App() {
       <Routes>
 
 
+        <Route path="/" element={<Login />} />
+
+
+
+
         {/* Admin Layout */}
         <Route path="/admin" element={<LayoutAdmin />} >
           {/* Outlet */}
-          <Route path='/admin' element={<LoginAdmin />} />
           <Route path='/admin/users' element={<UsersList />} />
+          <Route path='/admin/foods' element={<FoodsList />} />
           {/* ******** */}
         </Route>
+
+
+
+
+
+        {/* User Layout */}
+        {/* <Route path="/user" element={<LayoutAdmin />} > */}
+        {/* Outlet */}
+         <Route path='/user/foods' element={<FoodsList />} />
+
+
+        {/* ******** */}
+        {/* </Route> */}
+
+
+
 
       </Routes>
     </Router>
