+++ src/comps_admin/foods/foodsList.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function FoodsList() {
+  return (
+    <div>FoodsList</div>
+  )
+}
